Move email rendering into separate class.
