+++ tests/test_app/templates/email/html/html.php
@@ -1,5 +1,5 @@
 <?php
-use Cake\Mailer\Email;
+use Cake\Mailer\Renderer;
 ?>
 <h1>HTML Ipsum Presents</h1><p><strong>Pellentesque habitant morbi tristique</strong> senectus et netus et malesuada fames ac turpis egestas. Vestibulum tortor quam, feugiat vitae, ultricies eget, tempor sit amet, ante. Donec eu libero sit amet quam egestas semper. <em>Aenean ultricies mi vitae est.</em> Mauris placerat eleifend leo. Quisque sit amet est et sapien ullamcorper pharetra. Vestibulum erat wisi, condimentum sed, <code>commodo vitae</code>, ornare sit amet, wisi. Aenean fermentum, elit eget tincidunt condimentum, eros ipsum rutrum orci, sagittis tempus lacus enim ac dui. <a href="#">Donec non enim</a> in turpis pulvinar facilisis. Ut felis.</p><h2>Header Level 2</h2><ol><li>Lorem ipsum dolor sit amet, consectetuer adipiscing elit.</li><li>Aliquam tincidunt mauris eu risus.</li></ol><blockquote><p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus magna. Cras in mi at felis aliquet congue. Ut a est eget ligula molestie gravida. Curabitur massa. Donec eleifend, libero at sagittis mollis, tellus est malesuada tellus, at luctus turpis elit sit amet quam. Vivamus pretium ornare est.</p></blockquote><h3>Header Level 3</h3><ul><li>Lorem ipsum dolor sit amet, consectetuer adipiscing elit.</li><li>Aliquam tincidunt mauris eu risus.</li></ul>
 <pre><code>
@@ -10,6 +10,6 @@ use Cake\Mailer\Email;
 }</code></pre>
 <table>
 <th align="right" valign="top"
-        style="font-weight: bold">The tag is across multiple lines - <?= str_repeat('x', Email::LINE_LENGTH_MUST); ?></th>
+        style="font-weight: bold">The tag is across multiple lines - <?= str_repeat('x', Renderer::LINE_LENGTH_MUST); ?></th>
 </table>
-<p>Some more <?= str_repeat('x', Email::LINE_LENGTH_MUST); ?> <b>Bold</b> test.</p>
+<p>Some more <?= str_repeat('x', Renderer::LINE_LENGTH_MUST); ?> <b>Bold</b> test.</p>
